Refactor ContextManager
